requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
httpx
typing_extensions
